+++ Core/Src/Menu/MenuItem.cpp
@@ -1,18 +1,18 @@
 #include "MenuItem.h"
 
-MenuItem::MenuItem(int id, std::string title) {
+MenuItem::MenuItem(uint16_t id, std::string title) {
   this->title = title;
   this->id = id;
 }
 
-MenuItem::MenuItem(int id, std::string title, void (*callback)()) : MenuItem(id, title) {
+MenuItem::MenuItem(uint16_t id, std::string title, void (*callback)()) : MenuItem(id, title) {
   this->callback = callback;
   this->id = id;
 }
 
 std::string MenuItem::getTitle() { return this->title; }
 
-int MenuItem::getId() { return this->id; }
+uint16_t MenuItem::getId() { return this->id; }
 
 void MenuItem::executeCallback() { callback(); }
